webapi/tct-workers-w3c-tests/workers/w3c/Worker_ErrorEvent_type.htm

<!DOCTYPE html>
<html>
<head>
<title> AbstractWorker ErrorEvent.type </title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>
<script>
    setup({timeout:2000});
    var ErrorMessage = "Error Message";
    var FileName = './support/ErrorEvent.js';
    var description = "ErrorEvent object and ErrorEvent.type";
    
    var t = async_test("Test Description: " + description);
                  
    var worker = new Worker(FileName);

    worker.onmessage = t.step_func(function(evt)
    {
        assert_array_equals([evt.toString(), evt.type], ["[object MessageEvent]", "message"]);
        t.done();        
    });
    
    worker.postMessage(ErrorMessage);
</script>
</body>
</html>
